requirements.txt
flask
flask_wtf
flask_login
flask_sqlalchemy
sqlalchemy_migrate
